ref: apply andr macros to vault, primitive, and app contract
