+++ contracts/data-storage/andromeda-primitive/Cargo.toml
@@ -30,7 +30,6 @@ testing = ["cw-multi-test", "andromeda-testing"]
 cosmwasm-std = { workspace = true }
 cosmwasm-schema = { workspace = true }
 cw-storage-plus = { workspace = true }
-cw-utils = { workspace = true }
 cw20 = { workspace = true }
 
 andromeda-std = { workspace = true, features = ["rates"] }
