+++ contracts/ecosystem/andromeda-vault/Cargo.toml
@@ -15,8 +15,6 @@ cosmwasm-std = { workspace = true }
 cosmwasm-schema = { workspace = true }
 cw-storage-plus = { workspace = true }
 
-cw-utils = { workspace = true }
-
 andromeda-ecosystem = { workspace = true }
 andromeda-std = { workspace = true }
